[refactor] 내비게이션 수정
